requirements.txt
openai
anthropic
google-generativeai
chardet
tqdm 
tkinterdnd2